+++ src/components/StateApp/stateApp.js
@@ -1,3 +1,4 @@
 import React from 'react';
 
 export const ConvertContext = React.createContext({});
+export const ContextExchangeTabs = React.createContext({});
